+++ lez/indexer/service/protocol/src/lib.rs
@@ -48,7 +48,8 @@ mod base64 {
     }
 }
 
-// Bounds the server work one range query can request.
+// Largest block span a single events range query may cover. Lives here so every surface
+// that serves the query (RPC service, FFI) enforces the identical bound.
 pub const MAX_EVENT_QUERY_BLOCK_SPAN: u64 = 1000;
 
 pub type Nonce = u128;
@@ -230,7 +231,7 @@ pub struct PublicMessage {
     pub instruction_data: InstructionData,
 }
 
-pub type InstructionData = Vec<u32>;
+pub type InstructionData = Vec<u8>;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema)]
 pub struct PublicActionWithID {
